00.03_añadir_empleado_middelware_diseño

+++ app/Http/Controllers/EmpleadoController.php
@@ -3,26 +3,37 @@
 namespace App\Http\Controllers;
 
 use App\Models\Empleado;
-use Illuminate\Http\Request;
+use App\Models\Municipio;
+use App\Models\Provincia;
 use Illuminate\Support\Facades\Auth;
 
 class EmpleadoController extends Controller
 {
     public function listar()
     {
+        $provincias = Provincia::all();
+        $municipios = Municipio::all();
+
         $empleados = Empleado::where('id_empresa', Auth::user()->empresa_id)
             ->whereNull('deleted_at')
             ->orderBy('nombre', 'desc')
             ->paginate(5);
 
-        return view('empleado.lista', compact('empleados'));
+        return view('empleado.lista', ['empleados' => $empleados, 'provincias' => $provincias, 'municipios' => $municipios]);
     }
 
-    public function borrar($id_empleado)
+    public function borrar($id)
     {
-        $empleado = Empleado::where('id_empleado', $id_empleado);
+        $empleado = Empleado::where('id_empleado', $id);
+        $empleadoDatos = Empleado::where('id_empleado', $id)->first();
+
+        if ($empleadoDatos->id_empresa != Auth::user()->empresa_id) {
+            session()->flash('error', 'No tienes permiso para eliminar a este empleado.');
+            return redirect()->route('listarEmpleados');
+        }
+
         $empleado->delete();
-        session()->flash('message', 'El empleado ha sido borrado correctamente.');
+        session()->flash('message', "$empleadoDatos->nombre $empleadoDatos->apellidos ha sido dado de baja correctamente.");
         return redirect()->route('listarEmpleados');
     }
 }
